Bulk Change [2024-10-02]
Script sync from SHA: 36dbb1f0ccabc7fa91697f065d7315fcb4a9d3bf

+++ nvvkhl/gltf_scene_vk.cpp
@@ -115,7 +115,7 @@ static nvvkhl_shaders::GltfShadeMaterial getShaderMaterial(const tinygltf::Mater
   dstMat.pbrRoughnessFactor          = static_cast<float>(srcMat.pbrMetallicRoughness.roughnessFactor);
   dstMat.alphaMode         = srcMat.alphaMode == "OPAQUE" ? 0 : (srcMat.alphaMode == "MASK" ? 1 : 2 /*BLEND*/);
   dstMat.alphaCutoff       = static_cast<float>(srcMat.alphaCutoff);
-  dstMat.occlusionStrength = srcMat.occlusionTexture.strength;
+  dstMat.occlusionStrength = static_cast<float>(srcMat.occlusionTexture.strength);
   dstMat.occlusionTexture  = getTextureInfo(srcMat.occlusionTexture);
 
   KHR_materials_transmission transmission = tinygltf::utils::getTransmission(srcMat);
